fix: paginated fetching Commits case sensitivity
The Github API does not enforce casing on organization
login id. It will respond with the same data for "Netflix" or "netflix".
However, Apollo will cache these responses under different
keys, since the query is not exactly identical.

The original user entered "searchText" was not stored.
Thus the "login" stored in "organizationData" was supplied to
the "fetchMore" function. This caused an edge case where
the case of login id's might differ resulting in a cache miss
for paginated fetches. This prevented incoming data from
merging with cached data.

If a variable is not supplied in "fetchMore" function, then
the original value is used thereby avoiding this edge case.